update profile, add date to matches

+++ backend/controllers/profile.js
@@ -1,6 +1,8 @@
 /*CONTROLLER FOR PROFILE */
 import bcrypt, { compare } from "bcrypt"; 
 import User from "../models/User.js"; 
+import UserScoreModel from "../models/UserScoreModel.js"; // Import the UserScoreModel
+import PredictionModel from "../models/PredictionModel.js"; 
 
 
 export const getProfile = async (req, res) => {
@@ -72,24 +74,36 @@ export const updateProfile = async(req, res) =>{
     }
 }
 
-export const deleteProfile = async(req, res) =>{
-    try{
-        const userId = req.user.id; 
-        const user = await User.findById(userId); 
+export const deleteProfile = async (req, res) => {
+    try {
+        const userId = req.user.id;
+        const user = await User.findById(userId);
 
-        if(!user){
-            res.status(404).json({ message: "User not found"}); 
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
         }
 
-        const correctPassword = bcrypt.compare(req.body.currentPassword, user.password); 
-        if(!correctPassword){
-            return res.status(400).json({msg: "Password is incorrect"}); 
+        const password = req.body.password;
+        if (!password) {
+            return res.status(400).json({ message: "Password is required" });
         }
 
-        await User.findByIdAndDelete(userId); 
-        res.status(200).json({ message: "User account deleted successfully."}); 
+        const correctPassword = await bcrypt.compare(password, user.password);
+        if (!correctPassword) {
+            return res.status(400).json({ msg: "Password is incorrect" });
+        }
 
+        // Delete the user
+        await User.findByIdAndDelete(userId);
+
+        // Delete the associated UserScoreModel entry
+        await UserScoreModel.deleteMany({ userId });
+
+        // Delete the associated PredictionModel entries
+        await PredictionModel.deleteMany({ userId });
+
+        return res.status(200).json({ message: "User account and associated data deleted successfully." });
     } catch (err) {
-        res.status(500).json({ message: err.message}); 
+        return res.status(500).json({ message: err.message });
     }
-}
+};
